pyproject.toml
[project]
name = "databoard"
version = "0.1.0"
description = "Upload your data"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.112.2",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "polars>=1.33.1",
    "reflex>=0.8.11",
    "requests>=2.32.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.6.2",
    "fastapi-cli>=0.0.5",
]
